Allow to choose which version to use (Linux or Windows) if a game has both platform 1st try - i think something better can be done. Frustating about the dialog box used.

+++ minigalaxy/installer.py
@@ -15,7 +15,7 @@ def install_game(game, installer, parent_window=None) -> None:
         GLib.idle_add(__show_installation_error, game, _("{} failed to download.").format(installer), parent_window)
         raise FileNotFoundError("The installer {} does not exist".format(installer))
 
-    if game.platform == "linux":
+    if game.platform == "linux" and Config.get("platform") == "linux":
         if not __verify_installer_integrity(installer):
             GLib.idle_add(__show_installation_error, game, _("{} was corrupted. Please download it again.").format(installer), parent_window)
             os.remove(installer)
